Update com.gradle.plugin-publish to 0.11.0 (#19)
Description
===========

Due to a security issue with the old versions of the
`com.gradle.plugin-publish` plugin, all version < `0.11.0` won't
function anymore. We need to update to the latest version to be able to
publish plugins. Because this plugin uses itself to publish, I needed to
add the plugin with the newer version directly.

Changes
=======

* ![UPDATE] `com.gradle.plugin-publish` to `0.11.0`