Add release notes plugin (#34)
## Description
Added the github-release-notes plugin to generate more meaningful release notes. These notes are based in the included PRs `##Changes` section, and in lack of these, it uses the PRs names. There are references to the user which submited the PR and for the PR that generated these changes as well.