Add SonarQube plugin (#52)
Added the SonarQube gradle plugin to this project source code. The projects' source code and test coverage are sent to SonarQube when the sonarqube task is called, with the sonar.login property informing a valid SonarQube token and the sonar.projectKey property informing the sonar project key. This token can be generated at the user settings on SonarQube.

Eg.
./gradlew sonarqube -Dsonar.login=<<token>>

More details about the plugin in https://plugins.gradle.org/plugin/org.sonarqube

Co-authored-by: Joaquim Neto <joaquim.neto@wooga.net>