Add sonar defaults from github (#54)
Added defaults for sonarqube sonar.projectKey and sonar.projectName properties, based on github repository company and name. projectName is the same as the repository name, and projectKey is in the format ${company}_${repositoryName}. These fields can still be overwritten through environment variables or properties set through command line or gradle file.

For instance, for the github repository wooga/atlas-plugins we would have:

    ProjectKey: wooga_atlas-plugins
    ProjectName: atlas-plugins

This PR also includes a refactor for the sonar integration, as it was getting too complex to be kept inside the PluginsPlugin class. The sonarqube package was created, and now the configurations are built through the SonarQubeConfiguration class and still set in the PluginsPlugin.

Co-authored-by: Joaquim Neto <joaquim.neto@wooga.net>