Force groovy modules to version of version shipped with gradle (#65)
Description
===========

We have constant issues with groovy dependencies from different
libraries. One way of dealing with this in gradle is to add a force
resolution override. That worked well in the last couple of month but
has to be done in each project. To make this easier I deciced to pull
this into our base plugin for all plugins so we don't need to worry
about the version of gradle in use.

I'm still uncertain if this is a good idea though as I feel that we will
end up with other error especially when thinking of upgrading to gradle 7
which ships and works with groovy 3 which is in no way compatible to
groovy 2.5.
This setup I propose here seems to work. It should be easy enough be removed
if we see issues with it.

Changes
=======

* ![IMPROVE] groovy force override to use local groovy version from gradle