Improve plugin development with java-gradle-plugin
Description
===========

The [`java-gradle-plugin`] is the defactor standart how to
develop gradle plugins. It takes off a lot of work which
we had to maintain in our skeleton template projects.

It does basic validation and generates the plugin descriptor resource
property file automatically. The biggest benefit of using this
plugin is the support for composite builds which our current
system doesn't support.

The plugin can be tied together with the [`com.gradle.plugin-publish`] plugin.
Because the configuration needs upgrades I see this as a breaking change.

I updated the `build.gradle` file to follow the new setup and deleted the
no longer needed property file along with the resources sourceset.

Changes
=======

* ![ADD] `java-gradle-plugin`
* ![CHANGE] plugin publish config in `build.gradle`
* ![REMOVE] plugin descriptor property file

[`java-gradle-plugin`]: https://docs.gradle.org/current/userguide/java_gradle_plugin.html
[`com.gradle.plugin-publish`]: https://docs.gradle.org/current/userguide/publishing_gradle_plugins.html#plugin-publishing-plugin