Configuration avoidance API and branch name inference (#75)
Added configuration avoidance API use, as it is necessary for the project itself to not require github connection with the new implementation. It also has the added bonus of making gradle project synching faster.

Besides this, now we are using the new atlas-github version for git branch name inference. It first tries to get a property (github.branch.name), and if it does not exists, it goes with the following:

* if the current branch name starts with "PR-" (eg. PR-145), it fetches the PR branch name using github API.
* else, it returns the current local branch tracking branch name (as got by grgit)
* else, it returns the current local branch name (as got by grgit)

Co-authored-by: Joaquim Neto <joaquim.neto@wooga.net>
Co-authored-by: Manfred Endres <2523575+Larusso@users.noreply.github.com>