Update wooga dependencies to new group names and version
Description
===========

To publish our plugins with the `java-gradle-plugin` plugin means
a group change during publish. Since the plugins plugin is
using wooga plugins this update process is slighly circular.
In normal condition this doesn't really matter but in this case
it is needed. So I updated the `github`, `github-release-notes` and
the `version` plugin to their new groupname and version

Changes
=======

* ![UPDATE] `net.wooga.github` plugin groupname and version
* ![UPDATE] `net.wooga.github-release-notes` plugin groupname and version
* ![UPDATE] `net.wooga.version` plugin groupname and version